Renamed ReportCard and split up components.
